# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "operationId": "UsersController_findAll_api/v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "operationId": "UsersController_findById_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/products": {
      "post": {
        "operationId": "ProductsController_create_api/v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductsEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Products"
        ]
      },
      "get": {
        "operationId": "ProductsController_findAll_api/v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductsEntity"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
    "/api/v1/products/{id}": {
      "get": {
        "operationId": "ProductsController_findOneById_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductsEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Products"
        ]
      },
      "patch": {
        "operationId": "ProductsController_update_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductsEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_delete_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductsEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Products"
        ]
      }
    },
    "/api/v1/products/slug/{slug}": {
      "get": {
        "operationId": "ProductsController_findOneBySlug_api/v1",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductsEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
    "/api/v1/products/find": {
      "post": {
        "operationId": "ProductsController_findByIds_api/v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductsEntity"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
...
    "/api/v1/product-categories": {
      "post": {
        "operationId": "ProductCategoriesController_create_api/v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductCategoriesDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductCategoriesEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Product Categories"
        ]
      },
      "get": {
        "operationId": "ProductCategoriesController_findAll_api/v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductCategoriesEntity"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Product Categories"
        ]
      }
    },
    "/api/v1/product-categories/{id}": {
      "get": {
        "operationId": "ProductCategoriesController_findOne_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductCategoriesEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Product Categories"
        ]
      },
      "patch": {
        "operationId": "ProductCategoriesController_update_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductCategoriesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductCategoriesEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Product Categories"
        ]
      }
    },
...
  },
  "info": {
    "title": "Custom Admin Panel API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
      "CreateProductsDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "position": {
            "type": "integer",
            "format": "int32"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "social_media_links": {
            "type": "object",
            "nullable": true
          },
          "title": {
            "type": "string"
          },
          "short_description": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "category_id": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "item_weight_primary": {
            "type": "integer",
            "format": "int32"
          },
          "item_weight_secondary": {
            "type": "integer",
            "format": "int32"
          },
          "item_length": {
            "type": "integer",
            "format": "int32"
          },
          "item_width": {
            "type": "integer",
            "format": "int32"
          },
          "item_height": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "is_coming_soon": {
            "type": "boolean",
            "default": false
          }
        },
        "required": [
          "slug",
          "position",
          "images",
          "title",
          "short_description",
          "quantity",
          "item_weight_primary",
          "item_weight_secondary",
          "item_length",
          "item_width",
          "item_height",
          "price"
        ]
      },
      "ProductsEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "position": {
            "type": "integer",
            "format": "int32"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "social_media_links": {
            "type": "object",
            "nullable": true
          },
          "title": {
            "type": "string"
          },
          "short_description": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "category_id": {
            "type": "string",
            "nullable": true
          },
          "product_categories": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ProductCategoriesEntity"
              }
            ]
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "item_weight_primary": {
            "type": "integer",
            "format": "int32"
          },
          "item_weight_secondary": {
            "type": "integer",
            "format": "int32"
          },
          "item_length": {
            "type": "integer",
            "format": "int32"
          },
          "item_width": {
            "type": "integer",
            "format": "int32"
          },
          "item_height": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "is_new": {
            "type": "boolean"
          },
          "is_available": {
            "type": "boolean"
          },
          "is_frozen": {
            "type": "boolean"
          },
          "is_coming_soon": {
            "type": "boolean"
          },
          "is_deleted": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "slug",
          "position",
          "images",
          "social_media_links",
          "title",
          "short_description",
          "description",
          "category_id",
          "quantity",
          "item_weight_primary",
          "item_weight_secondary",
          "item_length",
          "item_width",
          "item_height",
          "price",
          "is_new",
          "is_available",
          "is_frozen",
          "is_coming_soon",
          "is_deleted",
          "created_at",
          "updated_at"
        ]
      },
      "ProductCategoriesEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "position": {
            "type": "integer",
            "format": "int32"
          },
          "image_url": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductsEntity"
            }
          }
        },
        "required": [
          "id",
          "title",
          "position",
          "image_url",
          "created_at",
          "updated_at"
        ]
      },
      "UpdateProductsDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "position": {
            "type": "integer",
            "format": "int32"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "social_media_links": {
            "type": "object",
            "nullable": true
          },
          "title": {
            "type": "string"
          },
          "short_description": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "category_id": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "item_weight_primary": {
            "type": "integer",
            "format": "int32"
          },
          "item_weight_secondary": {
            "type": "integer",
            "format": "int32"
          },
          "item_length": {
            "type": "integer",
            "format": "int32"
          },
          "item_width": {
            "type": "integer",
            "format": "int32"
          },
          "item_height": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "is_coming_soon": {
            "type": "boolean",
            "default": false
          }
        }
      },
...
      "CreateProductCategoriesDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "position": {
            "type": "integer",
            "format": "int32"
          },
          "image_url": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "position",
          "image_url"
        ]
      },
      "UpdateProductCategoriesDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "position": {
            "type": "integer",
            "format": "int32"
          },
          "image_url": {
            "type": "string"
          }
        }
      },
...
    }
  }
}